Total rewrite of the file structure and keybinds logic
Moved everything that was in user in separare folders: core for essentials, plugins for Lazy's configuration, keybinds for all the mappings and filetypes for shared settings between ftplugins.
Added a "Utils" file with shared functions.
Rewritten all the keybind logic to use shared functions and be declared with tables.
Moved all plugin keymappings to a single file for easier consulting while rewriting (this choice it might not be final).